Moved the files into separate folders nd fixed navigation and broken links
