+++ src/Dfe.RegionalImprovementForStandardsAndExcellence/Pages/Notes/Index.cshtml
@@ -12,7 +12,7 @@
 
 @section BeforeMain
 {
-    <a asp-page="@Model.ReturnPage"  class="govuk-back-link">@Links.ProjectList.Index.BackText</a>
+    <a asp-page="@Model.ReturnPage"  class="govuk-back-link">@Links.SchoolList.Index.BackText</a>
 }
 
 @if (Model.NewNote || Model.EditNote)
