+++ src/Dfe.RegionalImprovementForStandardsAndExcellence/Pages/Public/CookiePreferences.cshtml
@@ -150,7 +150,7 @@
       <h2 class="govuk-heading-m">Analytical cookies</h2>
 
       <h3 class="govuk-heading-s">Google Analytics</h3> 
-        <p class="govuk-body">Google Analytics sets the following cookies for <a asp-page="@Links.ProjectList.Index.Page">RISE</a>:</p>
+        <p class="govuk-body">Google Analytics sets the following cookies for <a asp-page="@Links.SchoolList.Index.Page">RISE</a>:</p>
       <table class="govuk-table" aria-label="Google analytics cookies">
          <thead class="govuk-table__header">
          <tr>
