+++ src/Dfe.RegionalImprovementForStandardsAndExcellence/Pages/SchoolList/Index.cshtml
@@ -11,7 +11,7 @@
         <h1 class="govuk-heading-xl" data-cy="select-heading">
             Schools requiring improvement
         </h1>
-        <a asp-page="/WhichSchoolNeedsHelp" role="button" draggable="false" class="govuk-button" data-module="govuk-button" data-cy="create_new_conversion_btn">
+        <a asp-page="/AddSchool/WhichSchoolNeedsHelp" role="button" draggable="false" class="govuk-button" data-module="govuk-button" data-cy="create_new_conversion_btn">
             Add a school
         </a>
     </div>
