+++ src/Dfe.RegionalImprovementForStandardsAndExcellence/Pages/TaskList/Index.cshtml
@@ -3,13 +3,13 @@
 @using Dfe.RegionalImprovementForStandardsAndExcellence.Frontend.TagHelpers
 @model Dfe.RegionalImprovementForStandardsAndExcellence.Frontend.Pages.TaskList.IndexModel
 @{
-    Model.SetErrorPage(Links.ProjectList.Index.Page);
+    Model.SetErrorPage(Links.SchoolList.Index.Page);
     ViewData["Title"] = "Task List";
 }
 
 @section BeforeMain
 {
-   <a asp-page="@Model.ReturnPage"  class="govuk-back-link">@Links.ProjectList.Index.BackText</a>
+   <a asp-page="@Model.ReturnPage"  class="govuk-back-link">@Links.SchoolList.Index.BackText</a>
 }
